requirements.txt
appdirs
attrs
backcall
black
Brotli
cairocffi
CairoSVG
cffi
click
cssselect2
cycler
dash
dash-canvas
dash-core-components
dash-html-components
dash-renderer
dash-table
decorator
defusedxml
Flask
Flask-Compress
future
gunicorn
imageio
importlib-metadata
ipython
ipython-genutils
itsdangerous
jedi
Jinja2
joblib
jsonschema
jupyter-core
kiwisolver
MarkupSafe
matplotlib
nbformat
networkx
nibabel
nilearn
numpy
packaging
pandas
parso
pathspec
pexpect
pickleshare
Pillow
plotly
prompt-toolkit
ptyprocess
pycparser
Pygments
pyparsing
pyrsistent
python-dateutil
pytz
PyWavelets
regex
retrying
scikit-image
scikit-learn
scipy
six
scikit-learn
threadpoolctl
tifffile
tinycss2
toml
traitlets
typed-ast
wcwidth
webencodings
Werkzeug
zipp
dash
Pillow
numpy
scipy
opencv-python
opencv-contrib-python
higra
matplotlib
utils
pydicom
ttkthemes

